Lots of changes

+++ src/Interfaces/RecordReader440.java
@@ -5,7 +5,7 @@ public interface RecordReader440<key, value> {
 	public void close();
 	public key createKey();
 	public value createValue();
-	public long getPos();
+	public int getPos();
 	public float getProgress();
 	public boolean next(key k, value v);
 }
